Patient, Dispensary, Role, & Staff CRUD Done | Status Class added | Security Config added

+++ src/main/java/com/odyssey/Ayurveda_Management/repository/PatientRepository.java
@@ -3,5 +3,5 @@ package com.odyssey.Ayurveda_Management.repository;
 import com.odyssey.Ayurveda_Management.model.Patient;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatientRepository extends JpaRepository<Patient, Long> {
+public interface PatientRepository extends JpaRepository<Patient, Integer> {
 }
